correct logic in annotation
